refactor(forgesvc): move provision/adopt/seed/resync seam to projectsync_provision.go (#963)
Pure same-package code motion, move-only (PRD #963 M3). Moves the board
provisioning surface — Adopt, Provision, provisionColors/provisionColor,
doneColumnName/doneColumnColor, appendDoneOption, provisionPrepare, the
seedParams type, seed/launchSeed/prepareSeedLink, adoptPrepare, Resync,
resyncPrepare, AutoCreateColumns, seedItems, Disable — plus the provision-only
helper unmatchedNote, verbatim (bodies and doc comments byte-identical) out of
projectsync.go into a new sibling file projectsync_provision.go, same package
forgesvc.

unmatchedNote caller map re-derived at this base: still called only from
adoptPrepare and resyncPrepare. projectSyncResolve, projectSyncPreamble,
ProjectSyncStatus (type+method), ensureProjectScope, optionMarker and
truncateErr stay in projectsync.go (now ~357 lines). No exported-API change
(go doc -all byte-identical to the M1 baseline), no behavior change, no test
edit; imports settled by the compiler. Verified byte-identical motion (850
moved declaration lines diff-clean) and go build/vet green.
